// src/lib/aiu/tables.gen.ts
export type ColumnType = "TEXT" | "BIGINT" | "DATE" | "TIMESTAMPTZ" | "JSONB";
export type TableName = "user_details" | "user_account_information" | "user_address_details" | "user_personal_details" | "client_details";